+++ src/djtools/rekordbox/__init__.py
@@ -2,14 +2,14 @@
 databases i.e. rekordbox.xml files.
 
 Included in this package:
+    * playlist_builder.py: Automatically create a desired playlist structure
+        based on the tags present in an XML.
     * randomize_tracks.py: Set tags of tracks in playlists sequentially
         (after shuffling) to randomize.
-    * rekordbox_playlist_builder.py: Automatically create a desired playlist
-        structure based on the tags present in an XML.
 """
 
-from .randomize_tracks import randomize_tracks
-from .rekordbox_playlist_builder import rekordbox_playlists
+from djtools.rekordbox.playlist_builder import rekordbox_playlists
+from djtools.rekordbox.randomize_tracks import randomize_tracks
 
 
 REKORDBOX_OPERATIONS = {
