+++ src/djtools/rekordbox/playlist_builder.py
@@ -19,47 +19,11 @@ from typing import Dict, List, Set, Tuple, Union
 import bs4
 from bs4 import BeautifulSoup
 
-from . import tag_parsers
-from .tag_parsers import Combiner
+from djtools.rekordbox import tag_parsers
+from djtools.rekordbox.tag_parsers import Combiner
 
 
 logger = logging.getLogger(__name__)
-
-
-def rekordbox_playlists(
-    config: Dict[str, Union[List, Dict, str, bool, int, float]]
-):
-    """Runs the PlaylistBuilder.
-
-    Args:
-        config: Configuration object.
-
-    Raises:
-        KeyError: "XML_PATH" must be set in "config.json".
-    """
-    rekordbox_database = config.get("XML_PATH")
-    if not rekordbox_database:
-        raise KeyError(
-            "Using the rekordbox_playlist_builder module requires the config "
-            "option XML_PATH"
-        ) from KeyError
-
-    playlist_config = os.path.join(
-        os.path.dirname(os.path.dirname(__file__)),
-        "configs",
-        "rekordbox_playlists.json",
-    ).replace(os.sep, "/")
-
-    pure_genre_playlists = config.get("GENRE_PLAYLISTS_PURE", [])
-    playlist_remainder_type = config.get("REKORDBOX_PLAYLISTS_REMAINDER")
-
-    playlist_builder = PlaylistBuilder(
-        rekordbox_database=rekordbox_database,
-        playlist_config=playlist_config,
-        pure_genre_playlists=pure_genre_playlists,
-        playlist_remainder_type=playlist_remainder_type,
-    )
-    playlist_builder()
 
 
 class PlaylistBuilder:
@@ -241,66 +205,6 @@ class PlaylistBuilder:
         ) as _file:
             _file.write(self._database.prettify("utf-8"))
 
-    def _create_playlists(
-        self,
-        soup: BeautifulSoup,
-        content: Union[str, Dict],
-        tags: Set[str] = set(),
-        top_level: bool = False,
-    ) -> bs4.element.Tag:
-        """Recursively traverses "rekordbox_playlists.json" and creates the
-            corresponding XML tag structure to be populated with tracks. If a
-            folder is encountered, an additional playlist is created called
-            "All <folder name>" (this does not apply to the top-level folder). If a
-            folder named "_ignore" is encountered, tag playlists specified in the
-            associated "playlists" list will not have a tag created for them, but
-            the tags will be added to the "My Tags" set so the corresponding
-            tracks will be ignored when generating the "Other" folder / playlist.
-
-        Args:
-            soup: Parsed XML.
-            content: Playlist name or folder name with playlists.
-            tags: Playlist "My Tags" set to populate.
-            top_level: Flag used to indicate that no "All" playlist should
-                be created at the top-level of the playlist tree.
-
-        Raises:
-            ValueError: "rekordbox_playlists.json" must be properly formatted.
-
-        Returns:
-            Populated playlist structure.
-        """
-        if isinstance(content, dict):
-            content = {k.lower(): v for k, v in content.items()}
-            if content["name"] == "_ignore":
-                tags.update(set(content["playlists"]))
-            else:
-                folder = soup.new_tag("NODE", Name=content["name"], Type="0")
-                if not top_level:
-                    _all = soup.new_tag(
-                        "NODE",
-                        KeyType="0",
-                        Name=f'All {content["name"]}',
-                        Type="1",
-                    )
-                    folder.append(_all)
-                for playlist in content["playlists"]:
-                    _playlist = self._create_playlists(
-                        soup=soup, content=playlist, tags=tags
-                    )
-                    if _playlist:
-                        folder.append(_playlist)
-                return folder
-        elif isinstance(content, str):
-            tags.add(content)
-            playlist = soup.new_tag("NODE", KeyType="0", Name=content, Type="1")
-            return playlist
-        else:
-            raise ValueError(
-                f"Encountered invalid input type {type(content)}: {content}"
-            )
-
-
     def _add_other(
         self,
         soup: BeautifulSoup,
@@ -337,7 +241,6 @@ class PlaylistBuilder:
             playlists.append(playlist)
         else:
             logger.error(f'Invalid remainder type "{remainder_type}"')
-
 
     def _add_tracks(
         self,
@@ -439,3 +342,97 @@ class PlaylistBuilder:
                         break
                     parent = parent.parent
             
+    def _create_playlists(
+        self,
+        soup: BeautifulSoup,
+        content: Union[str, Dict],
+        tags: Set[str] = set(),
+        top_level: bool = False,
+    ) -> bs4.element.Tag:
+        """Recursively traverses "rekordbox_playlists.json" and creates the
+            corresponding XML tag structure to be populated with tracks. If a
+            folder is encountered, an additional playlist is created called
+            "All <folder name>" (this does not apply to the top-level folder). If a
+            folder named "_ignore" is encountered, tag playlists specified in the
+            associated "playlists" list will not have a tag created for them, but
+            the tags will be added to the "My Tags" set so the corresponding
+            tracks will be ignored when generating the "Other" folder / playlist.
+
+        Args:
+            soup: Parsed XML.
+            content: Playlist name or folder name with playlists.
+            tags: Playlist "My Tags" set to populate.
+            top_level: Flag used to indicate that no "All" playlist should
+                be created at the top-level of the playlist tree.
+
+        Raises:
+            ValueError: "rekordbox_playlists.json" must be properly formatted.
+
+        Returns:
+            Populated playlist structure.
+        """
+        if isinstance(content, dict):
+            content = {k.lower(): v for k, v in content.items()}
+            if content["name"] == "_ignore":
+                tags.update(set(content["playlists"]))
+            else:
+                folder = soup.new_tag("NODE", Name=content["name"], Type="0")
+                if not top_level:
+                    _all = soup.new_tag(
+                        "NODE",
+                        KeyType="0",
+                        Name=f'All {content["name"]}',
+                        Type="1",
+                    )
+                    folder.append(_all)
+                for playlist in content["playlists"]:
+                    _playlist = self._create_playlists(
+                        soup=soup, content=playlist, tags=tags
+                    )
+                    if _playlist:
+                        folder.append(_playlist)
+                return folder
+        elif isinstance(content, str):
+            tags.add(content)
+            playlist = soup.new_tag("NODE", KeyType="0", Name=content, Type="1")
+            return playlist
+        else:
+            raise ValueError(
+                f"Encountered invalid input type {type(content)}: {content}"
+            )
+
+
+def rekordbox_playlists(
+    config: Dict[str, Union[List, Dict, str, bool, int, float]]
+):
+    """Runs the PlaylistBuilder.
+
+    Args:
+        config: Configuration object.
+
+    Raises:
+        KeyError: "XML_PATH" must be set in "config.json".
+    """
+    rekordbox_database = config.get("XML_PATH")
+    if not rekordbox_database:
+        raise KeyError(
+            "Using the rekordbox.playlist_builder module requires the config "
+            "option XML_PATH"
+        ) from KeyError
+
+    playlist_config = os.path.join(
+        os.path.dirname(os.path.dirname(__file__)),
+        "configs",
+        "rekordbox_playlists.json",
+    ).replace(os.sep, "/")
+
+    pure_genre_playlists = config.get("GENRE_PLAYLISTS_PURE", [])
+    playlist_remainder_type = config.get("REKORDBOX_PLAYLISTS_REMAINDER")
+
+    playlist_builder = PlaylistBuilder(
+        rekordbox_database=rekordbox_database,
+        playlist_config=playlist_config,
+        pure_genre_playlists=pure_genre_playlists,
+        playlist_remainder_type=playlist_remainder_type,
+    )
+    playlist_builder()
