+++ src/djtools/rekordbox/test_tag_parsers.py
@@ -14,75 +14,6 @@ pytest_plugins = [
 ]
 
 
-def test_tagparser_raises_type_error():
-    with pytest.raises(
-        TypeError,
-        match=(
-            "Can't instantiate abstract class TagParser with abstract method "
-            "__call__"
-        ),
-    ):
-        TagParser()
-
-
-def test_tagparser_call_raises_not_imlemented_error(test_track):
-    TagParser.__abstractmethods__ = set()
-    class TagParserSubclass(TagParser):
-        pass
-
-    with pytest.raises(
-        NotImplementedError,
-        match=(
-            "Classes inheriting from TagParser must override the __call__ "
-            "method."
-        ),
-    ):
-        TagParserSubclass(parser_config={})(test_track)
-
-
-def test_genretagparser(test_playlist_config, test_xml):
-    with open(test_playlist_config, mode="r", encoding="utf-8",) as _file:
-        playlist_config = json.load(_file)
-    genre = "Techno"
-    parser = GenreTagParser(
-        parser_config=playlist_config["GenreTagParser"],
-        pure_genre_playlists=[genre],
-    )
-    with open(test_xml, mode="r", encoding="utf-8") as _file:
-        db = BeautifulSoup(_file.read(), "xml")
-    tracks = {
-        track["TrackID"]: track for track in db.find_all("TRACK")
-        if track.get("Location")
-    }
-    playlist = db.find_all("NODE", {"Name": genre, "Type": "1"})[0]
-    for track_key in playlist.find_all("TRACK"):
-        track = tracks[track_key["Key"]]
-        tags = parser(track)
-        assert genre in tags
-        if f"Pure {genre}" in tags:
-            assert all(
-                genre in tag for tag in tags
-            )
-
-
-def test_mytagparser(test_playlist_config, test_xml):
-    with open(test_playlist_config, mode="r", encoding="utf-8",) as _file:
-        playlist_config = json.load(_file)
-    mytag = "Hypnotic"
-    parser = MyTagParser(parser_config=playlist_config["MyTagParser"])
-    with open(test_xml, mode="r", encoding="utf-8") as _file:
-        db = BeautifulSoup(_file.read(), "xml")
-    tracks = {
-        track["TrackID"]: track for track in db.find_all("TRACK")
-        if track.get("Location")
-    }
-    playlist = db.find_all("NODE", {"Name": mytag, "Type": "1"})[0]
-    for track_key in playlist.find_all("TRACK"):
-        track = tracks[track_key["Key"]]
-        tags = parser(track)
-        assert mytag in tags
-
-
 @pytest.mark.parametrize(
     "node_attributes",
     [
@@ -133,6 +64,32 @@ def test_booleannode_raises_runtime_eror():
         node({})
 
 
+def test_tagparser_raises_type_error():
+    with pytest.raises(
+        TypeError,
+        match=(
+            "Can't instantiate abstract class TagParser with abstract method "
+            "__call__"
+        ),
+    ):
+        TagParser()
+
+
+def test_tagparser_call_raises_not_imlemented_error(test_track):
+    TagParser.__abstractmethods__ = set()
+    class TagParserSubclass(TagParser):
+        pass
+
+    with pytest.raises(
+        NotImplementedError,
+        match=(
+            "Classes inheriting from TagParser must override the __call__ "
+            "method."
+        ),
+    ):
+        TagParserSubclass(parser_config={})(test_track)
+
+
 def test_combiner(test_playlist_config, test_xml, caplog):
     with open(test_playlist_config, mode="r", encoding="utf-8",) as _file:
         playlist_config = json.load(_file)
@@ -176,3 +133,46 @@ def test_combiner_raises_lookuperror_for_bad_playlist(
             parser_config=playlist_config["Combiner"],
             rekordbox_database=db,
         )
+
+
+def test_genretagparser(test_playlist_config, test_xml):
+    with open(test_playlist_config, mode="r", encoding="utf-8",) as _file:
+        playlist_config = json.load(_file)
+    genre = "Techno"
+    parser = GenreTagParser(
+        parser_config=playlist_config["GenreTagParser"],
+        pure_genre_playlists=[genre],
+    )
+    with open(test_xml, mode="r", encoding="utf-8") as _file:
+        db = BeautifulSoup(_file.read(), "xml")
+    tracks = {
+        track["TrackID"]: track for track in db.find_all("TRACK")
+        if track.get("Location")
+    }
+    playlist = db.find_all("NODE", {"Name": genre, "Type": "1"})[0]
+    for track_key in playlist.find_all("TRACK"):
+        track = tracks[track_key["Key"]]
+        tags = parser(track)
+        assert genre in tags
+        if f"Pure {genre}" in tags:
+            assert all(
+                genre in tag for tag in tags
+            )
+
+
+def test_mytagparser(test_playlist_config, test_xml):
+    with open(test_playlist_config, mode="r", encoding="utf-8",) as _file:
+        playlist_config = json.load(_file)
+    mytag = "Hypnotic"
+    parser = MyTagParser(parser_config=playlist_config["MyTagParser"])
+    with open(test_xml, mode="r", encoding="utf-8") as _file:
+        db = BeautifulSoup(_file.read(), "xml")
+    tracks = {
+        track["TrackID"]: track for track in db.find_all("TRACK")
+        if track.get("Location")
+    }
+    playlist = db.find_all("NODE", {"Name": mytag, "Type": "1"})[0]
+    for track_key in playlist.find_all("TRACK"):
+        track = tracks[track_key["Key"]]
+        tags = parser(track)
+        assert mytag in tags
