+++ src/djtools/spotify/README.md
@@ -7,30 +7,34 @@
 
 # Overview
 The `spotify` package contains modules:
-* `spotify_playlist_builder`: constructs / updates Spotify playlists from subreddit posts or the Discord webhook output from music uploads
+* `playlist_builder`: constructs / updates Spotify playlists from subreddit posts or the Discord webhook output from music uploads
 
 # Setup
 To begin using the `spotify` package, you must create a Spotify API application by following [these instructions](https://developer.spotify.com/documentation/web-api/quick-start/). Once you've registered your application, you must populate the `SPOTIFY_CLIENT_ID`, `SPOTIFY_CLIENT_SECRET`, and `SPOTIFY_REDIRECT_URI` configuration options in `config.json`.
 
-If using the `spotify_playlist_builder` module's `update_auto_playlists` funciton, you must also create a Reddit API application by following [these instructions](https://rymur.github.io/setup). Once you've registered your application, you must populate the `REDDIT_CLIENT_ID`, `REDDIT_CLIENT_SECRET`, and `REDDIT_USER_AGENT` configuration options in `config.json`.
+If using the `playlist_builder` module's `update_auto_playlists` function, you must also create a Reddit API application by following [these instructions](https://rymur.github.io/setup). Once you've registered your application, you must populate the `REDDIT_CLIENT_ID`, `REDDIT_CLIENT_SECRET`, and `REDDIT_USER_AGENT` configuration options in `config.json`.
 
-If using either the `update_auto_playlists` of `playlist_from_upload` function of `spotify_playlist_builder`, `SPOTIFY_USERNAME` must be set to the user who will keep the playlists on their account.
+If using either the `update_auto_playlists` or `playlist_from_upload` function of `playlist_builder`, `SPOTIFY_USERNAME` must be set to the user who will keep the playlists on their account.
 
 # Usage
 
-## spotify_playlist_builder
-In order to use the `spotify_playlist_builder` module, you must add a `spotify_playlists.json` file to the `config` folder. This JSON contains playlist names as keys and Spotify playlist IDs as values. If the playlist names are used in the `update_auto_playlists` function, the JSON keys _MUST_ match a valid subreddit (case-insensitive) because these keys are used to query the Reddit API. Playlist names for the `playlist_from_upload` function are automatically generated using the usernames present in the Discord webhook output.
+## playlist_builder
+In order to use the `playlist_builder` module, you must add a `spotify_playlists.json` file to the `config` folder. This JSON contains playlist names as keys and Spotify playlist IDs as values. If the playlist names are used in the `update_auto_playlists` function, the JSON keys _MUST_ match a valid subreddit (case-insensitive) because these keys are used to query the Reddit API. Playlist names for the `playlist_from_upload` function are automatically generated using the usernames present in the Discord webhook output. `spotify_playlists.json` may also contain other playlists not used by either the `playlist_from_upload` or `update_auto_playlists` functions (i.e. for use with `utils.compare_tracks`).
 
-`NOTE`: Top-level folders inside of "DJ Music" should be the usernames of Beatcloud users; the Spotify playlists generated by the `playlists_from_upload` function will be named "<folder_name> Uploads".
+`NOTE`: Top-level folders inside of "DJ Music" should be the usernames of Beatcloud users; the Spotify playlists generated by the `playlists_from_upload` function will be named "<folder_name> Uploads". If the names of these playlists are changed from  the "<folder_name> Uploads" format or the top-level folders of "DJ Music" are not usernames, then you will experience unexpected results when using `playlist_from_upload` or `sync.download_music` with the `DOWNLOAD_INCLUDE_SPOTIFY` option.
 
-For the `update_auto_playlists` function, you must add the playlist names you want to generate / update to the `AUTO_PLAYLIST_SUBREDDITS` configuration option. `AUTO_PLAYLIST_FUZZ_RATIO` `[0, 100]` sets the minimum Levenshtein similarity to add tracks to an auto-playlist when comparing the subreddit post title to a Spotify API search result. This configuration is also used by the `playlist_from_upload` function to compare Spotify search results with file names present in the Discord webhook output from a music upload. `AUTO_PLAYLIST_SUBREDDIT_LIMIT` `[0, 999]` restricts the number of subreddit posts to consider when attempting to search Spotify for tracks. If this option isn't specified, then a value of `500` will be used. If the option's value is falsey, then `None` will be used i.e. there will be no limit and 999 posts will be retrieved (if possible). 
+For the `update_auto_playlists` function, you must add configuration objects for the subreddit playlist names you want to generate / update to the `AUTO_PLAYLIST_SUBREDDITS` configuration option.
 
-Each element of `AUTO_PLAYLIST_SUBREDDITS` is a dictionary of subreddit-specific configuration options for the `spotify_playlist_builder` module:
+Each element of `AUTO_PLAYLIST_SUBREDDITS` is a dictionary of subreddit-specific configuration options for the `update_auto_playlists` function:
 * `name` is the subreddit name.
 * `type` is whether to query subreddit's "hot", "top", "new", "rising", or "controversial" posts
-* `period` is the time period for which posts are considered (`week` is recommended); should a subreddit post not contain a direct link to a Spotify track
+* `period` is the time period for which posts are considered (`week` is recommended)
 * `limit` is the maximum number of tracks in the auto-playlist; the oldest track will be removed when adding a new track exceeds this limit
 
-Triggering `spotify_playlist_builder.update_auto_playlists` can be done by setting `AUTO_PLAYLIST_UPDATE: true`.
+`AUTO_PLAYLIST_FUZZ_RATIO` `[0, 100]` sets the minimum Levenshtein similarity to add tracks to an auto-playlist when comparing the subreddit post title to a Spotify API search result. The `CHECK_TRACK_OVERLAP_FUZZ_RATIO` configuration option is used by the `playlist_from_upload` function to compare Spotify search results with file names present in the Discord webhook output from a music upload.
 
-Triggering `spotify_playlist_builder.playlist_from_upload` can be done by setting `PLAYLIST_FROM_UPLOAD: true` (Discord webhook output must be copied to the system clipboard). Alternatively, the Discord webhook output can be saved to a file and that file can be passed to the configuration option as `PLAYLIST_FROM_UPLOAD: "path/to/file.txt"`.
+`AUTO_PLAYLIST_SUBREDDIT_LIMIT` `[0, 999]` restricts the number of subreddit posts to consider when attempting to search Spotify for tracks. If this option isn't specified, then a value of `500` will be used. If the option's value is falsey, then `None` will be used i.e. there will be no limit and 999 posts will be retrieved (if possible). 
+
+Triggering `playlist_builder.update_auto_playlists` can be done by setting `AUTO_PLAYLIST_UPDATE: true`.
+
+Triggering `playlist_builder.playlist_from_upload` can be done by setting `PLAYLIST_FROM_UPLOAD: true` (Discord webhook output must be copied to the system clipboard). Alternatively, the Discord webhook output can be saved to a file and that file can be passed to the configuration option as `PLAYLIST_FROM_UPLOAD: "path/to/file.txt"`.
