+++ src/djtools/sync/README.md
@@ -7,10 +7,10 @@
 
 # Overview
 The `sync` package contains a module, `sync_operation`, which can:
-* `upload_music`: sync files from your `USB_PATH` to the `beatcloud` which do not yet exit at the destination
-* `download_music`: sync files from the `beatcloud` to your `USB_PATH` which do not yet exit at the destination
+* `download_music`: sync files from the `beatcloud` to your `USB_PATH`
+* `download_xml`: sync the XML file belonging to `XML_IMPORT_USER` from their folder under `xml` in the `beatcloud` to the parent folder of `XML_PATH`. This operation also rewrites the downloaded XML replacing the track Location fields of `XML_IMPORT_USER` with `USB_PATH` by looking up the `USB_PATH` of `XML_IMPORT_USER` (according to `registered_users.json`).
+* `upload_music`: sync files from your `USB_PATH` to the `beatcloud`
 * `upload_xml`: sync your `XML_PATH` to your `USER` folder under the `xml` folder in the `beatcloud`
-* `download_xml`: sync the `XML_IMPORT_USER` XML from their folder under the `xml` folder in the `beatcloud` to the folder where your `XML_PATH` is...this operation also rewrites the `XML_IMPORT_USER` XML so that the track locations point to those in your `USB_PATH` by looking up the `USB_PATH` that `XML_IMPORT_USER` has set (according to `registered_users.json`) and replaces it with your own `USB_PATH`.
 
 
 # Setup
@@ -22,11 +22,11 @@ The setup required for all these operations is having `awscli` installed and con
 
 `upload_xml` requires that `XML_PATH` exists.
 
-`download_xml` requires that the folder where `XML_PATH` lives exists and that `XML_IMPORT_USER` has run `upload_xml`; it's also required that `XML_IMPORT_USER` has an entry in `registered_users.json`...this will be the case if they have ever run `dj_tools.py` with a valid `config.json` and then pushed their changes to `registered_users.json` prior to your most recent pull of `DJ Tools`.
+`download_xml` requires that the folder where `XML_PATH` lives exists and that `XML_IMPORT_USER` has run `upload_xml`; it's also required that `XML_IMPORT_USER` has an entry in `registered_users.json` which maps their username i.e `XML_IMPORT_USER` to whatever drive they have set as `USB_PATH`.
 
 # Usage
 To run any of the four sync operations {`DOWNLOAD_XML`, `DOWNLOAD_MUSIC`, `UPLOAD_XML`, `UPLOAD_MUSIC`}, just set the corresponding config option to true. If running `DOWNLOAD_XML` make sure `XML_IMPORT_USER` is set to the proper `beatcloud` user. Before uploading / after downloading, you can follow any of the prescribed workflows specified under the main README's "Basic Information" section to ensure consistency both in your local library and in the `beatcloud`.
 
-If the `DOWNLOAD_INCLUDE_SPOTIFY` options is set to the name of a Spotify playlist that exists in your `spotify_playlists.json` config, only tracks appearing in that playlist will be downloaded from the Beatcloud.
+If the `DOWNLOAD_INCLUDE_SPOTIFY` options is set to the name of a Spotify playlist that exists in your `spotify_playlists.json` config and follows the format `<folder_name> Uploads` where `<folder_name>` matches a top-level folder of "DJ Music" in the beatcloud (ideally should be a user's name), only tracks appearing in that playlist will be downloaded from the Beatcloud.
 
-`NOTE`: you can run `UPLOAD_MUSIC` and / or `DOWNLOAD_MUSIC` with the `--dryrun` flag to inspect the result of the same sync operations without actually performing them.
+`NOTE`: you can run `UPLOAD_MUSIC` and / or `DOWNLOAD_MUSIC` with the `--dryrun` flag to inspect the result of the sync operations without actually performing them.
