+++ src/djtools/sync/__init__.py
@@ -1,9 +1,8 @@
 """The 'sync' package contains a module for syncing tracks to and from the
 beatcloud, syncing USER's XML to the beatcloud, and syncing XML_IMPORT_USER's
-XML to the parent directory of 'XML_PATH'. It also contains a module of helper
-functions used by the former module.
+XML to the parent directory of 'XML_PATH'.
 """
-from .sync_operations import (
+from djtools.sync.sync_operations import (
     download_music, download_xml, upload_music, upload_xml
 )
