+++ src/djtools/sync/sync_operations.py
@@ -9,89 +9,14 @@ import os
 from pathlib import Path
 from typing import Dict, List, Union
 
-from djtools.sync.helpers import parse_sync_command, rewrite_xml, run_sync, \
-                                 webhook
-from djtools.utils.helpers import compare_tracks, make_dirs
+from djtools.sync.helpers import (
+    parse_sync_command, rewrite_xml, run_sync, webhook
+)
+from djtools.utils.check_track_overlap import compare_tracks
+from djtools.utils.helpers import make_dirs
 
 
 logger = logging.getLogger(__name__)
-
-
-def upload_music(config: Dict[str, Union[List, Dict, str, bool, int, float]]):
-    """This function syncs tracks from "USB_PATH" to the beatcloud.
-        "AWS_USE_DATE_MODIFIED" can be used in order to reupload tracks that
-        already exist in the beatcloud but have been modified since the last
-        time they were uploaded (i.e. ID3 tags have been altered).
-
-    Args:
-        config: Configuration object.
-
-    Raises:
-        KeyError: "USB_PATH" must be configured.
-        FileNotFoundError: "USB_PATH" must exist.
-    """
-    try:
-        usb_path = config["USB_PATH"]
-    except KeyError:
-        raise KeyError(
-            "Using the upload_music function of the sync_operations module "
-            "requires the config option USB_PATH"
-        ) from KeyError
-
-    if not os.path.exists(usb_path):
-        raise FileNotFoundError(f'USB_PATH "{usb_path}" does not exist!')
-
-    glob_path = Path(os.path.join(usb_path, "DJ Music").replace(os.sep, "/"))
-    hidden_files = {
-        str(p) for p in glob_path.rglob(
-            os.path.join("**", ".*.*").replace(os.sep, "/")
-        )
-    }
-    if hidden_files:
-        logger.info(f"Removed {len(hidden_files)} files...")
-        for _file in hidden_files:
-            logger.info(f"\t{_file}")
-            os.remove(_file)
-
-    logger.info("Syncing track collection...")
-    src = os.path.join(usb_path, "DJ Music").replace(os.sep, "/")
-    cmd = ["aws", "s3", "sync", src, "s3://dj.beatcloud.com/dj/music/"]
-
-    if config.get("DISCORD_URL") and not config.get("DRYRUN"):
-        webhook(
-            config["DISCORD_URL"],
-            content=run_sync(parse_sync_command(cmd, config, upload=True)),
-        )
-    else:
-        run_sync(parse_sync_command(cmd, config, upload=True))
-
-
-def upload_xml(config: Dict[str, Union[List, Dict, str, bool, int, float]]):
-    """This function uploads "XML_PATH" to beatcloud.
-
-    Args:
-        config: Configuration object.
-
-    Raises:
-        KeyError: "XML_PATH" must be configured.
-        FileNotFoundError: "XML_PATH" file must exist.
-    """
-    try:
-        xml_path = config["XML_PATH"]
-    except KeyError:
-        raise KeyError(
-            "Using the upload_xml function of the sync_operations module "
-            "requires the config option XML_PATH"
-        ) from KeyError
-
-    if not os.path.exists(xml_path):
-        raise FileNotFoundError(f'XML_PATH "{xml_path}" does not exist!')
-
-    logger.info(f"Uploading {config['USER']}'s rekordbox.xml...")
-    dst = f's3://dj.beatcloud.com/dj/xml/{config["USER"]}/'
-    cmd = f"aws s3 cp {xml_path} {dst}"
-    logger.info(cmd)
-    os.system(cmd)
 
 
 def download_music(
@@ -210,3 +135,80 @@ def download_xml(config: Dict[str, Union[List, Dict, str, bool, int, float]]):
     logger.info(cmd)
     os.system(cmd)
     rewrite_xml(config)
+
+
+def upload_music(config: Dict[str, Union[List, Dict, str, bool, int, float]]):
+    """This function syncs tracks from "USB_PATH" to the beatcloud.
+        "AWS_USE_DATE_MODIFIED" can be used in order to reupload tracks that
+        already exist in the beatcloud but have been modified since the last
+        time they were uploaded (i.e. ID3 tags have been altered).
+
+    Args:
+        config: Configuration object.
+
+    Raises:
+        KeyError: "USB_PATH" must be configured.
+        FileNotFoundError: "USB_PATH" must exist.
+    """
+    try:
+        usb_path = config["USB_PATH"]
+    except KeyError:
+        raise KeyError(
+            "Using the upload_music function of the sync_operations module "
+            "requires the config option USB_PATH"
+        ) from KeyError
+
+    if not os.path.exists(usb_path):
+        raise FileNotFoundError(f'USB_PATH "{usb_path}" does not exist!')
+
+    glob_path = Path(os.path.join(usb_path, "DJ Music").replace(os.sep, "/"))
+    hidden_files = {
+        str(p) for p in glob_path.rglob(
+            os.path.join("**", ".*.*").replace(os.sep, "/")
+        )
+    }
+    if hidden_files:
+        logger.info(f"Removed {len(hidden_files)} files...")
+        for _file in hidden_files:
+            logger.info(f"\t{_file}")
+            os.remove(_file)
+
+    logger.info("Syncing track collection...")
+    src = os.path.join(usb_path, "DJ Music").replace(os.sep, "/")
+    cmd = ["aws", "s3", "sync", src, "s3://dj.beatcloud.com/dj/music/"]
+
+    if config.get("DISCORD_URL") and not config.get("DRYRUN"):
+        webhook(
+            config["DISCORD_URL"],
+            content=run_sync(parse_sync_command(cmd, config, upload=True)),
+        )
+    else:
+        run_sync(parse_sync_command(cmd, config, upload=True))
+
+
+def upload_xml(config: Dict[str, Union[List, Dict, str, bool, int, float]]):
+    """This function uploads "XML_PATH" to beatcloud.
+
+    Args:
+        config: Configuration object.
+
+    Raises:
+        KeyError: "XML_PATH" must be configured.
+        FileNotFoundError: "XML_PATH" file must exist.
+    """
+    try:
+        xml_path = config["XML_PATH"]
+    except KeyError:
+        raise KeyError(
+            "Using the upload_xml function of the sync_operations module "
+            "requires the config option XML_PATH"
+        ) from KeyError
+
+    if not os.path.exists(xml_path):
+        raise FileNotFoundError(f'XML_PATH "{xml_path}" does not exist!')
+
+    logger.info(f"Uploading {config['USER']}'s rekordbox.xml...")
+    dst = f's3://dj.beatcloud.com/dj/xml/{config["USER"]}/'
+    cmd = f"aws s3 cp {xml_path} {dst}"
+    logger.info(cmd)
+    os.system(cmd)
